# services/backend/docs/openapi/backend.yaml
openapi: 3.1.0
info:
  title: Gift Suggestion Backend API
  version: 0.2.0
  description: |
    Backend API gateway for the Gift Suggestion Web Service.
    The specification is maintained manually and currently covers health,
    auth/user foundation, catalog read endpoints, wishlist endpoints,
    admin catalog import endpoints, recommendation endpoints, tracking endpoints,
    and VK integration scaffold endpoints.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/wishlists:
    get:
      tags: [wishlist]
      summary: List current user wishlists (deprecated)
      description: Deprecated compatibility endpoint. Returns zero or one wishlist for the current user.
      operationId: listWishlistsCompat
      deprecated: true
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Wishlist list page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WishlistListEnvelope'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [wishlist]
      summary: Create current user wishlist (deprecated)
      description: Deprecated compatibility endpoint. Creates the single personal wishlist if it does not exist yet.
      operationId: createWishlistCompat
      deprecated: true
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWishlistRequest'
      responses:
        '201':
          description: Wishlist created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WishlistEnvelope'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Current user already has a wishlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/wishlist:
    get:
      tags: [wishlist]
      summary: Get my wishlist
      operationId: getCurrentWishlist
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Personal wishlist detail. Created automatically on first access when missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WishlistEnvelope'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [wishlist]
      summary: Delete my wishlist
      operationId: deleteCurrentWishlist
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Personal wishlist deleted. If it does not exist, the operation is still successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteWishlistEnvelope'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/wishlist/items:
    post:
      tags: [wishlist]
      summary: Add gift to my wishlist
      operationId: addCurrentWishlistItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddWishlistItemRequest'
      responses:
        '200':
          description: Gift is already saved in wishlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddWishlistItemEnvelope'
        '201':
          description: Gift added to wishlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddWishlistItemEnvelope'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Gift not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/wishlist/items/{gift_id}:
    delete:
      tags: [wishlist]
      summary: Remove gift from my wishlist
      operationId: removeCurrentWishlistItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GiftIDPath'
      responses:
        '200':
          description: Gift removed from wishlist. If it is absent, the operation is still successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoveWishlistItemEnvelope'
        '400':
          description: Invalid identifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/wishlists/{wishlist_id}:
    get:
      tags: [wishlist]
      summary: Get current user wishlist with items (deprecated)
      description: Deprecated compatibility endpoint. The path wishlist_id must match the current user's single wishlist.
      operationId: getWishlistCompat
      deprecated: true
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WishlistIDPath'
      responses:
        '200':
          description: Wishlist detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WishlistEnvelope'
        '400':
          description: Invalid identifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Wishlist not found or unavailable to current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [wishlist]
      summary: Delete current user wishlist (deprecated)
      description: Deprecated compatibility endpoint. The path wishlist_id must match the current user's single wishlist.
      operationId: deleteWishlistCompat
      deprecated: true
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WishlistIDPath'
      responses:
        '200':
          description: Wishlist deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteWishlistEnvelope'
        '400':
          description: Invalid identifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Wishlist not found or unavailable to current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/wishlists/{wishlist_id}/items:
    post:
      tags: [wishlist]
      summary: Add gift to current user wishlist (deprecated)
      description: Deprecated compatibility endpoint. The path wishlist_id must match the current user's single wishlist.
      operationId: addWishlistItemCompat
      deprecated: true
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WishlistIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddWishlistItemRequest'
      responses:
        '201':
          description: Gift added to wishlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddWishlistItemEnvelope'
        '200':
          description: Gift is already saved in wishlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddWishlistItemEnvelope'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Wishlist or gift not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/wishlists/{wishlist_id}/items/{gift_id}:
    delete:
      tags: [wishlist]
      summary: Remove gift from current user wishlist (deprecated)
      description: Deprecated compatibility endpoint. The path wishlist_id must match the current user's single wishlist.
      operationId: removeWishlistItemCompat
      deprecated: true
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WishlistIDPath'
        - $ref: '#/components/parameters/GiftIDPath'
      responses:
        '200':
          description: Gift removed from wishlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoveWishlistItemEnvelope'
        '400':
          description: Invalid identifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    GiftIDPath:
      in: path
      name: gift_id
      required: true
      schema:
        type: string
        format: uuid
    WishlistIDPath:
      in: path
      name: wishlist_id
      required: true
      schema:
        type: string
        format: uuid
...
    LimitQuery:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    OffsetQuery:
      in: query
      name: offset
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
  schemas:
    Meta:
      type: object
      properties:
        request_id:
          type: string
      additionalProperties: false
    ErrorPayload:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
      additionalProperties: false
    ErrorEnvelope:
      type: object
      required: [status, error, meta]
      properties:
        status:
          type: string
          const: error
        error:
          $ref: '#/components/schemas/ErrorPayload'
        meta:
          $ref: '#/components/schemas/Meta'
      additionalProperties: false
...
    Page:
      type: object
      required: [limit, offset, total]
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
      additionalProperties: false
...
    CatalogGiftCategory:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      additionalProperties: false
    CatalogGift:
      type: object
      required: [id, name, description, price, store_link, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: string
          pattern: '^\d+(\.\d{2})$'
        store_link:
          type: string
          format: uri
        image:
          type: string
          format: uri
        age_restriction:
          type: integer
          enum: [0, 12, 16, 18]
        category:
          $ref: '#/components/schemas/CatalogGiftCategory'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
...
    CreateWishlistRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: Deprecated compatibility field. Personal wishlist name is normalized server-side.
      additionalProperties: false
    AddWishlistItemRequest:
      type: object
      required: [gift_id]
      properties:
        gift_id:
          type: string
          format: uuid
      additionalProperties: false
    WishlistItem:
      type: object
      required: [id, created_at, gift]
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        gift:
          $ref: '#/components/schemas/CatalogGift'
      additionalProperties: false
    WishlistSummary:
      type: object
      required: [id, name, item_count, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Personal wishlist display name.
        item_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
    Wishlist:
      type: object
      required: [id, name, item_count, items, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Personal wishlist display name.
        item_count:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/WishlistItem'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
    WishlistEnvelope:
      type: object
      required: [status, data, meta]
      properties:
        status:
          type: string
          const: ok
        data:
          type: object
          required: [wishlist]
          properties:
            wishlist:
              $ref: '#/components/schemas/Wishlist'
          additionalProperties: false
        meta:
          $ref: '#/components/schemas/Meta'
      additionalProperties: false
    WishlistListEnvelope:
      type: object
      required: [status, data, meta]
      properties:
        status:
          type: string
          const: ok
        data:
          type: object
          required: [items, page]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/WishlistSummary'
            page:
              $ref: '#/components/schemas/Page'
          additionalProperties: false
        meta:
          $ref: '#/components/schemas/Meta'
      additionalProperties: false
    AddWishlistItemEnvelope:
      type: object
      required: [status, data, meta]
      properties:
        status:
          type: string
          const: ok
        data:
          type: object
          required: [already_in_wishlist, item]
          properties:
            already_in_wishlist:
              type: boolean
            item:
              $ref: '#/components/schemas/WishlistItem'
          additionalProperties: false
        meta:
          $ref: '#/components/schemas/Meta'
      additionalProperties: false
    RemoveWishlistItemEnvelope:
      type: object
      required: [status, data, meta]
      properties:
        status:
          type: string
          const: ok
        data:
          type: object
          required: [removed]
          properties:
            removed:
              type: boolean
          additionalProperties: false
        meta:
          $ref: '#/components/schemas/Meta'
      additionalProperties: false
    DeleteWishlistEnvelope:
      type: object
      required: [status, data, meta]
      properties:
        status:
          type: string
          const: ok
        data:
          type: object
          required: [deleted]
          properties:
            deleted:
              type: boolean
          additionalProperties: false
        meta:
          $ref: '#/components/schemas/Meta'
      additionalProperties: false
...